refactor: moved constants and styling to separate files
