+++ src/components/cars-list-component/constants.ts
@@ -1,0 +1,1 @@
+export const PAGE_LIMIT = 4;
